msm: mdss: isolate cleanup of pipes within kickoff
During display commit cycle, we need to iterate through pipes that have
been marked for cleanup and destroy these. However in the middle of the
function we may block for sometime to wait for display to be ready, when
this happens we are releasing ov_lock which controls access to this
cleanup list. If there is any cleanup done in this period, then there
could be some pipes incorrectly cleaned up after this.

To fix this race condition, add a local variable to keep track of pipes
that should be cleaned up.

Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>
Signed-off-by: Raghavendra Ambadas <rambad@codeaurora.org>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_overlay.c

Change-Id: Ib80cf91e10d66e318754c327fd71a489b8c42676